Start liveness check endpoint in frontend
The frontend is the parent process, and the service itself the child processm
so start liveness check endpoint in the frontend and send to the service,
rather than the other way round.
